refactor: Store.OpenStatus 클래스 분리
