+++ common/src/main/java/com/food/common/store/domain/type/OpenStatus.java
@@ -1,0 +1,13 @@
+package com.food.common.store.domain.type;
+
+public enum OpenStatus {
+    OPEN("운영 중"),
+    CLOSED("운영 종료")
+    ;
+
+    private final String description;
+
+    OpenStatus(String description) {
+        this.description = description;
+    }
+}
