+++ common/src/test/java/com/food/common/mock/store/MockStore.java
@@ -3,6 +3,7 @@ package com.food.common.mock.store;
 import com.food.common.mock.user.MockUser;
 import com.food.common.store.domain.Store;
 import com.food.common.store.domain.StoreOwner;
+import com.food.common.store.domain.type.OpenStatus;
 
 public class MockStore {
     public static Builder builder()  {
@@ -18,7 +19,7 @@ public class MockStore {
                 .build())
                 .build();
         private Integer minOrderAmount = 10000;
-        private Store.OpenStatus status = Store.OpenStatus.OPEN;
+        private OpenStatus status = OpenStatus.OPEN;
 
         public Builder id(Long id) {
             this.id = id;
@@ -40,7 +41,7 @@ public class MockStore {
             return this;
         }
 
-        public Builder status(Store.OpenStatus status) {
+        public Builder status(OpenStatus status) {
             this.status = status;
             return this;
         }
